// frontend/src/api-client/models/PlaylistWithDetails.ts
export interface PlaylistWithDetails {
  /**
   *
   * @type {string}
   * @memberof PlaylistWithDetails
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof PlaylistWithDetails
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof PlaylistWithDetails
   */
  description: string | null;
  /**
   *
   * @type {string}
   * @memberof PlaylistWithDetails
   */
  creatorUserId: string;
  /**
   *
   * @type {Date}
   * @memberof PlaylistWithDetails
   */
  createdAt: Date;
  /**
   *
   * @type {Date}
   * @memberof PlaylistWithDetails
   */
  updatedAt: Date;
  /**
   *
   * @type {Array<PlaylistCollaborator>}
   * @memberof PlaylistWithDetails
   */
  collaborators: Array<PlaylistCollaborator>;
  /**
   *
   * @type {Array<PlaylistClip>}
   * @memberof PlaylistWithDetails
   */
  clips: Array<PlaylistClip>;
}